refactor: 이름 변경 Home -> Info
