+++ src/main/java/me/minkh/app/controller/InfoController.java
@@ -2,23 +2,23 @@ package me.minkh.app.controller;
 
 import lombok.extern.slf4j.Slf4j;
 import me.minkh.app.dto.info.InfoResponseDto;
-import me.minkh.app.service.HomeService;
+import me.minkh.app.service.info.InfoService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
 
 @Slf4j
 @RestController
-public class HomeController {
+public class InfoController {
 
-    private final HomeService homeService;
+    private final InfoService infoService;
 
-    public HomeController(HomeService homeService) {
-        this.homeService = homeService;
+    public InfoController(InfoService homeService) {
+        this.infoService = homeService;
     }
 
     @GetMapping("/info/{characterName}")
     public InfoResponseDto info(@PathVariable("characterName") String characterName) {
-        return this.homeService.info(characterName);
+        return this.infoService.info(characterName);
     }
 }
