+++ src/main/java/me/minkh/app/service/info/converter/ProfileDtoConverter.java
@@ -1,4 +1,4 @@
-package me.minkh.app.service.converter;
+package me.minkh.app.service.info.converter;
 
 import me.minkh.app.dto.lostark.ProfileDto;
 import me.minkh.app.dto.info.ProfileStatResponseDto;
